defect_detection/pyproject.toml
[project]
name = "defect-detection"
version = "2.0.0"
description = "Advanced YOLO-based defect detection system"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "ultralytics>=8.2.0",
  "typer>=0.12",
  "rich>=13.7",
  "PyYAML>=6.0",
  "opencv-python>=4.8",
  "numpy>=1.23",
  "pandas>=2.0",
  "scikit-learn>=1.3",
  "matplotlib>=3.8",
  "tqdm>=4.66"
]

[project.scripts]
defect-detect = "defect_detection.cli:app"

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
